requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.14
